requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0

